\name{tinythemes-package}
\alias{tinythemes-package}
\alias{tinythemes}
\docType{package}
\title{\packageTitle{tinythemes}}
\description{\packageDescription{tinythemes}}
\section{Package Content}{\packageIndices{tinythemes}}
\author{\packageAuthor{tinythemes}}
\section{Maintainer}{\packageMaintainer{tinythemes}}
\keyword{package}
